Rewrite blockchain class
